+++ CHANGELOG.md
@@ -4,11 +4,15 @@
 
 * IntuneAccountProtectionLocalAdministratorPasswordSolutionPolicy
   * Migrate to new settings catalog cmdlets.
+* IntuneSettingCatalogASRRulesPolicyWindows10
+  * Migrate to new settings catalog cmdlets.
 * IntuneExploitProtectionPolicyWindows10SettingCatalog
   * Migrate to new settings catalog cmdlets.
 * M365DSCDRGUtil
   * Fixes an issue with the settings catalog property generation.
   * Force array as parameter in `Compare-M365DSCIntunePolicyAssignment`.
+  * Fixes issues with values of type `groupSettingCollection` and `choiceSetting`  
+    when creating the settings catalog policy settings body.
 * M365DSCUtil
   * Fixes an issue where the comparison with null-valued desired value throws an error.